requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# AI and ML
google-generativeai==0.3.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1

# Data validation
marshmallow==3.20.1

# HTTP and API
httpx==0.25.0
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Date and time
python-dateutil==2.8.2

# Async support
asyncio-mqtt==0.13.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.0

# Development
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Build backend
setuptools>=68.0.0
wheel>=0.41.0

# Production
gunicorn==21.2.0
prometheus-client==0.18.0